requirements.txt
numpy~=1.26.4
matplotlib~=3.10.0
pandas~=2.2.3
seaborn~=0.13.0
pomdp-py~=1.3.5.1
torch~=2.6.0
scikit-learn~=1.6.1
pygame~=2.1.0
torchvision~=0.21.0
openpyxl~=3.1.2
opencv-python~=4.11.0.86
networkx~=3.3
pyyaml~=6.0
optuna~=4.2.1
plotly~=5.14.1
ray[tune]~=2.45.0
dm-tree~=0.1.9
hpbandster~=0.7.4
ConfigSpace~=1.2.0
lz4~=4.4.3
DI-engine[test,fast]
transformers~=4.50.2
python-sat~=1.8.dev13
